organized the file structure a little bit and created the ActionManager, class responsible for creating Action commands (GRASP creator pattern)
