+++ src/main/java/ca/mcmaster/se2aa4/island/teamXXX/drone/Direction.java
@@ -1,4 +1,4 @@
-package ca.mcmaster.se2aa4.island.teamXXX;
+package ca.mcmaster.se2aa4.island.teamXXX.drone;
 
 public enum Direction {
     NORTH("N"),
