+++ src/main/java/ca/mcmaster/se2aa4/island/teamXXX/drone/Position.java
@@ -1,4 +1,4 @@
-package ca.mcmaster.se2aa4.island.teamXXX;
+package ca.mcmaster.se2aa4.island.teamXXX.drone;
 
 public class Position {
     private int x;
